+++ src/api/workspace/end-point.constants.ts
@@ -1,3 +1,3 @@
-export const GNB_LEFT_API_ENDPOINTS = {
+export const WORKSPACE_API_ENDPOINTS = {
   RECENT_DOCUMENTS: (workspaceId: number | null) => `workspaces/${workspaceId}/sidebar?workspaceId`,
 } as const;
