+++ src/common/components/gnbs/GnbLeft/RecentDocumentsSection/RecentDocumentsSection.server.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import useFetchRecentDocuments from '@common/hooks/queries/gnb-left/useFetchRecentDocuments';
+import useFetchRecentDocuments from '@/api/workspace/get/queries/useFetchRecentDocuments';
 
 import RecentDocumentItem from '../RecentDocumentItem/RecentDocumentItem.server';
 import RecentDocumentItemSkeleton from '../RecentDocumentItem/RecentDocumentItemSkeleton.server';
