+++ src/features/on-boarding/apis/post/create-workspace.ts
@@ -1,6 +1,6 @@
 import { defaultApi } from '@lib/fetcher';
 
-import { BaseResponseDto } from '@common/types/api.types';
+import { BaseResponseDto } from '@common/types/api.common.types';
 
 import {
   CreateWorkspaceRequestDto,
